# perf(rome_formater): Inline `FormatElement::Text`