--- a/crates/rome_formatter/src/source_map.rs
+++ b/crates/rome_formatter/src/source_map.rs
@@ -206,8 +206,9 @@ impl TransformSourceMap {
     /// in the original, untransformed source code.
     ///
     /// The printer creates a source map that allows mapping positions from the newly formatted document
-    /// back to the locations of the tree. However, the source positions stored in [crate::FormatElement::Text]
-    /// are relative to the transformed tree and not the original tree passed to [crate::format_node].
+    /// back to the locations of the tree. However, the source positions stored in [crate::FormatElement::DynamicText]
+    /// and [crate::FormatElement::SyntaxTokenText] are relative to the transformed tree
+    /// and not the original tree passed to [crate::format_node].
     ///
     /// This function re-maps the positions from the positions in the transformed tree back to the positions
     /// in the original, untransformed tree.