--- a/crates/rome_formatter/src/source_map.rs
+++ b/crates/rome_formatter/src/source_map.rs
@@ -207,7 +207,7 @@ impl TransformSourceMap {
     ///
     /// The printer creates a source map that allows mapping positions from the newly formatted document
     /// back to the locations of the tree. However, the source positions stored in [crate::FormatElement::DynamicText]
-    /// and [crate::FormatElement::SyntaxTokenText] are relative to the transformed tree
+    /// and [crate::FormatElement::SyntaxTokenTextSlice] are relative to the transformed tree
     /// and not the original tree passed to [crate::format_node].
     ///
     /// This function re-maps the positions from the positions in the transformed tree back to the positions