## Summary

This PR inlines the `Text` enum into the `FormatElement` enum to reduce the size of `FormatElement` from 32 to 24 bytes.

The largest variant of `FormatElement` is `FormatElement::Text` with a size of 24 bytes, which makes `FormatElement` 32 bytes wide because it adds one byte for the variant tag + 7 bytes of padding. 

The size of `Text` is 24 bytes:
* 1 byte for the union tag.
* 20 bytes for the largest variant `SyntaxTokenText`: 
  * 16 bytes for the slice  (8 bytes for `GreenToken` + 8 bytes for the text range)
  * 4 bytes for the source position
* 3 bytes of padding

This means, that we reserve 10 bytes of padding even for the largest enum variant (even worse for e.g. `LineSuffixBoundary` that has 31 bytes of padding). 

This PR inlines `Text` to avoid having to pad twice, once for `Text` and once on `FormatElement` which results in a size reduction of `FormatElement` to 24 bytes (3 bytes of padding for `SyntaxTokenText`). 

The downside of this change is that matching on text becomes more complicated and it removes the `Eq` implementation that compared `Text` by the string content. 

## Test Plan

<details>
<summary>This change results in a 20% peak memory consumption reduction: Max Bytes: 87.29 MB down to 69.48 MB</summary>


```bash
cargo run -p xtask_bench --release --features=dhat-heap -- --feature=formatter  --criterion=false --filter tex
```

### Main

```
[tex-chtml-full.js] - using [/home/micha/git/rome/target/tex-chtml-full.js]
Start
Formatted
        Memory
                Current Blocks: 208143
                Current Bytes: 87.09 MB
                Max Blocks: 208146
                Max Bytes: 87.29 MB
                Total Blocks: 3689951
                Total Bytes: 370.88 MB
Printed
        Memory
                Current Blocks: 148775
                Current Bytes: 21.44 MB
                Max Blocks: 208150
                Max Bytes: 97.12 MB
                Total Blocks: 71
                Total Bytes: 20.06 MB
Benchmark: https://cdn.jsdelivr.net/npm/mathjax@3.2.0/es5/tex-chtml-full.js
        Formatting: 8.768215985s
                      ----------
        Total:        8.768215985s

Summary
-------
tex-chtml-full.js, Formatting: 8.768215985s
dhat: Total:     463,866,958 bytes in 3,870,874 blocks
dhat: At t-gmax: 101,839,473 bytes in 208,150 blocks
dhat: At t-end:  1,152 bytes in 3 blocks
dhat: The data has been saved to dhat-heap.json, and is viewable with dhat/dh_view.html
```

### This Branch

```
[tex-chtml-full.js] - using [/home/micha/git/rome/target/tex-chtml-full.js]
Start
Formatted
        Memory
                Current Blocks: 208143
                Current Bytes: 69.28 MB
                Max Blocks: 208146
                Max Bytes: 69.48 MB
                Total Blocks: 3689951
                Total Bytes: 329.41 MB
Printed
        Memory
                Current Blocks: 148775
                Current Bytes: 21.44 MB
                Max Blocks: 208150
                Max Bytes: 79.32 MB
                Total Blocks: 71
                Total Bytes: 20.06 MB
Benchmark: https://cdn.jsdelivr.net/npm/mathjax@3.2.0/es5/tex-chtml-full.js
        Formatting: 8.808443181s
                      ----------
        Total:        8.808443181s

Summary
-------
tex-chtml-full.js, Formatting: 8.808443181s
dhat: Total:     420,388,694 bytes in 3,870,874 blocks
dhat: At t-gmax: 83,169,561 bytes in 208,150 blocks
dhat: At t-end:  1,152 bytes in 3 blocks
dhat: The data has been saved to dhat-heap.json, and is viewable with dhat/dh_view.html
``` 

</details>

I expect performance to improve a bit because it reduces the amount of data written during formatting. However, I don't expect too big of an improvement because the size reduction isn't big enough to make 4 format elements fit into a single cache line of 64 bytes (it's now 2 2/3 which probably is of very little use)

Local perf results:

```
group                                    main                                    text
-----                                    ----                                    ----
formatter/checker.ts                     1.00    216.2±1.31ms    12.0 MB/sec     1.02   221.4±18.09ms    11.7 MB/sec
formatter/compiler.js                    1.10   133.0±12.24ms     7.9 MB/sec     1.00    120.6±8.90ms     8.7 MB/sec
formatter/d3.min.js                      1.00     92.1±1.55ms     2.8 MB/sec     1.05    96.5±10.32ms     2.7 MB/sec
formatter/dojo.js                        1.00      6.5±0.02ms    10.6 MB/sec     1.06      6.8±0.63ms    10.0 MB/sec
formatter/ios.d.ts                       1.08   145.6±19.56ms    12.8 MB/sec     1.00   134.9±14.83ms    13.8 MB/sec
formatter/jquery.min.js                  1.00     26.7±0.11ms     3.1 MB/sec     1.04     27.9±2.55ms     3.0 MB/sec
formatter/math.js                        1.04    196.6±9.30ms     3.3 MB/sec     1.00   189.0±15.78ms     3.4 MB/sec
formatter/parser.ts                      1.04      4.6±0.01ms    10.7 MB/sec     1.00      4.4±0.07ms    11.0 MB/sec
formatter/pixi.min.js                    1.09   112.7±11.09ms     3.9 MB/sec     1.00    103.6±8.16ms     4.2 MB/sec
formatter/react-dom.production.min.js    1.07     33.6±3.26ms     3.4 MB/sec     1.00     31.5±3.08ms     3.7 MB/sec
formatter/react.production.min.js        1.02  1533.8±24.99µs     4.0 MB/sec     1.00  1498.6±12.46µs     4.1 MB/sec
formatter/router.ts                      1.07      4.1±0.38ms    15.0 MB/sec     1.00      3.8±0.21ms    16.1 MB/sec
formatter/tex-chtml-full.js              1.10   265.7±23.36ms     3.4 MB/sec     1.00   240.8±20.58ms     3.8 MB/sec
formatter/three.min.js                   1.00    126.8±9.68ms     4.6 MB/sec     1.00   127.0±13.34ms     4.6 MB/sec
formatter/typescript.js                  1.01   873.1±69.30ms    10.9 MB/sec     1.00   866.7±77.22ms    11.0 MB/sec
formatter/vue.global.prod.js             1.05     45.6±4.58ms     2.6 MB/sec     1.00     43.4±4.41ms     2.8 MB/sec
```